Adds more contrast to button focus state
This is part of a wider effort to make the focus state of all form/interactive elements have more contrast. After accesibility reviews of a few new apps using Rivet it has come up that the focus states in Rivet are not critical, but are certainly a weakness from an accessibility perspective.